Avatar group limit.
